Move createProject method to TestBase
